Add Serial1 ESP-8266 UART to FleaFPGA-Uno
This adds a parameter to UARTClass to allow specifying the SIO register
to use (defaults to main SIO port so other boards should not be
affected).  It also adds a default Serial1 object for FleaFPGA-Uno
variant (which uses this dedicated UART to talk to optional ESP-8266
WiFi).  Allows easy support of as many UARTs as configured on a board.